# swagger_doc.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Banco de musicas",
    "description": "API REST com uso de sequelize mySQL, que gerencia usuarios, ADM e USER, para um banco de dados de musica"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/musicas/{limite}/{pagina}": {
      "get": {
        "summary": "usuario logado lista musicas por paginacao",
        "description": "",
        "parameters": [
          {
            "name": "limite",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pagina",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/bandas/{limite}/{pagina}": {
      "get": {
        "summary": "usuario logado lista bandas por paginacao",
        "description": "",
        "parameters": [
          {
            "name": "limite",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pagina",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/albuns/{limite}/{pagina}": {
      "get": {
        "summary": "usuario logado lista albuns por paginacao",
        "description": "",
        "parameters": [
          {
            "name": "limite",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pagina",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/musica/{id}": {
      "get": {
        "summary": "usuario logado procura musica por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/album/{id}": {
      "get": {
        "summary": "usuario logado procura album por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/album/{id}/musicas/{limite}/{pagina}": {
      "get": {
        "summary": "usuario logado procura musica por ID de album com paginacao",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "limite",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pagina",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/banda/{id}/musicas/{limite}/{pagina}": {
      "get": {
        "summary": "usuario logado procura musica por ID de banda com paginacao",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "limite",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pagina",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/banda/{id}/albuns/{limite}/{pagina}": {
      "get": {
        "summary": "usuario logado procura albuns por ID de banda com paginacao",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "limite",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "pagina",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}